Grade a native OLMo-core checkpoint generatively, without converting it
leaderboard_math needed an HF checkpoint and Preston's is native, so the
choice was to convert or to write the missing backend. Conversion is the
worse door: save_hf_model has to understand each submitter's architecture,
and this checkpoint already needed the get_hf_config patch because a plain
pre-norm block raises upstream. The native reader takes whatever
TransformerConfig.from_dict parses.

The work is an adapter, not new machinery. Everything between a prompt and
a graded response was already backend-agnostic -- the per-item budget, the
context clamp, the exemplar ladder, the EOS resolution and the stop list
all sit above the completer, and GenerativeScorer builds the fitter from
what a completer publishes. So a backend is a completer plus a row in the
registry, and _OlmoCoreCompleter publishes five attributes and decodes.

Three of those five fail silently when absent, which is why each is tested
by removal rather than by presence. Without a context window the fitter
returns unclamped, MATH's long stems keep four exemplars, and the model is
truncated into the block that teaches \boxed{} and Final Answer: -- zeros
that look like the model rather than the harness.

The window is read from the checkpoint rather than defaulted, and a
checkpoint that declares none is refused at load. _resolve_max_length was
deliberately not reused whole: its last fallback is the tokenizer's
model_max_length, and SmolLM2 declares 8192 against weights trained at
2048, which would leave the clamp on, inert, and four times too generous
-- worse than no clamp, because it looks resolved.

--dtype now reaches generation. The MCQ half of that flag was inert for a
whole run while the report said bfloat16, which is the failure this
prevents repeating.

Also corrects two comments in inference.py still carrying the pad/eos
objection that probe run_019fe316-0e47 withdrew: the validator rejects only
pad == eos, and fabricating the pad leaves the real EOS at 0.

+++ diagnostics/mcq_cat/runner.py
@@ -238,20 +238,26 @@ def run(args: argparse.Namespace) -> int:
         )
         return 0
 
-    # `--dtype` reaches two places from here, because there are two ways a precision can
-    # be chosen and the flag has to mean the same thing whichever one is live: below it
-    # goes to `prepare_checkpoint`, which writes converted weights at it, and here it
-    # goes to the config the native backend builds its model from. Exactly one of those
-    # applies per run -- conversion happens under `--checkpoint-prep auto`, the native
-    # loader runs under `none` -- so setting both is not a conflict, and setting only the
-    # first is what left the flag inert on the native path.
+    # `--dtype` reaches three places from here, because there are two ways a precision
+    # can be chosen and two modalities that can choose it, and the flag has to mean the
+    # same thing whichever is live: below it goes to `prepare_checkpoint`, which writes
+    # converted weights at it, and here it goes to both configs a native backend builds
+    # its model from. Exactly one of the three applies per run -- conversion happens
+    # under `--checkpoint-prep auto`, a native loader runs under `none`, and a run has
+    # one modality -- so setting all three is not a conflict, and setting only the first
+    # is what left the flag inert on the native path. The generative half was added with
+    # the native completer rather than after it, because the MCQ half spent a whole run
+    # reporting bfloat16 while scoring in float32 and nothing in the artifact said so.
     settings = grading.GradingSettings(
         mcq=inference.InferenceConfig(
             checkpoint_kind=args.checkpoint_kind,
             batch_size=args.batch_size,
             dtype=args.dtype,
         ),
-        generation=generative.GenerationConfig(checkpoint_kind=args.checkpoint_kind),
+        generation=generative.GenerationConfig(
+            checkpoint_kind=args.checkpoint_kind,
+            dtype=args.dtype,
+        ),
     )
 
     with tempfile.TemporaryDirectory(prefix="mcq-cat-") as tmp:
